Adding a couple npcs for new mall area. (#123)
The first one is an upgraded version of the crawler that gives better items, and more xp on kill but also does more damage.
The second one is an entirely new kind of npc: the walker security officer. This walker is unique because it is the first walker that wears body armor, but has no helmet to compensate.