Add "The Many" boss NPC (#143)
* Statting "The Many"

Drops will most definitely need to be changed and more will most likely need to be added to compensate for the lack of weapon this boss uses. However, I think this is a decent starting point for this boss. It's also the first boss to have a chance to bite players.

* Update src/resources/npcs.ts

Co-authored-by: blobfysh <47480729+blobfysh@users.noreply.github.com>

* [Linter] Apply linters fixes

* remove chance to bite

* temporary fix for melee boss

Co-authored-by: blobfysh <47480729+blobfysh@users.noreply.github.com>
Co-authored-by: EriisGay <EriisGay@users.noreply.github.com>